+++ ui/base/l10n/formatter.cc
@@ -13,6 +13,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 
 #include "base/check.h"
 #include "base/component_export.h"
+#include "base/numerics/safe_conversions.h"
 #include "third_party/icu/source/common/unicode/unistr.h"
 #include "third_party/icu/source/i18n/unicode/msgfmt.h"
 #include "ui/base/l10n/l10n_util.h"
@@ -284,7 +285,9 @@ std::unique_ptr<icu::MessageFormat> Formatter::InitFormat(
     std::u16string pattern = l10n_util::GetStringUTF16(pluralities.id);
     UErrorCode error = U_ZERO_ERROR;
     std::unique_ptr<icu::MessageFormat> format(new icu::MessageFormat(
-        icu::UnicodeString(false, pattern.data(), pattern.length()), error));
+        icu::UnicodeString(false, pattern.data(),
+                           base::checked_cast<int32_t>(pattern.length())),
+        error));
     DCHECK(U_SUCCESS(error));
     if (format.get())
       return format;
